requirements.txt
flask
python-pptx
pymupdf
requests
gunicorn
